Add files via upload
When adding a normal map, textures become more detailed and fit better to the shape of the object that is being textured.  When adding a displacement map on top of this, an illusion of depth of surface is created.  This is because part of the mesh now appear to be either raised or lowered from their neighbors, which causes an effect where the surface of the object has depth and feels lifelike.